refactor: split VRM component info into VRM0 / VRM1
