+++ Editor/APIInternal/ComponentInfos.VRM1.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: cda357a894554e8bb67cb0fd3cd02659
+timeCreated: 1698673843
